Command record may return cached records (bug #2799).
The 'record' command with argument 'offset' given but with no
'syntax' and 'esn' parameter now fetches records from last search.
This allows fast retrieval of marcxml and raw records. If, OTOH, esn
and syntax is given the record is assumed not to be cached and a new
present request is fired as before.