New mergekey facilities. Bug #2343 and bug #2800).
The mergekey (recid in display) is now generated by the following
rules: the contents of attribute mergekey and all metadata elements
with attribute mergekey=yes are combined in to one resulting mergekey.
If the resulting mergekey is empty a unique mergekey is generated
(which prevents merging). This allows a system with no merging
enabled. It also allows mergekeys to be customized "per-service".