Character normalization per-service.
The elements 'relevance', 'sort', 'mergekey' may now be given
inside service element as well as server. The server elements are
for a service if it does not specify the normalization itself.
The settings and normalization info is now inheried to each service
in a separate pass to ensure that order does not matter.