Fix connect timeout problem on Solaris.
This problem was seen on Solaris but could happen anywhere. Pazpar2
did not detect that Connection was established. And consequently did
not change timeout from 'Z39.50 connect timeout' to 'Z39.50 session
timeout'. Since The former is only 15 seconds by default a slow search
would be aborted when it shouldn't. This commit removes a call to
ZOOM_connection_process which, on Solaris, would grab one or more
events out, so that Pazpar2 was not informed.