Mergekey from XSL overrides mergekey=yes.
If mergekey attribute is present in XSL (px:xslt), then value for that
is used (and value of elements defined in metadata mergekey=yes are not
used). OIf mergekey is not present, then value of metadata mergekey=yes
elements are used. An empty value mergekey from XSL will generate
a unique mergekey.