Settings, bytarget + db out-of-range check
Command bytarget with argument settings=1 will show settings
per target.. This is to be able to verify correct settings and
be able to test that they are correct. The database settings
array size is now also stored..  Problems with database settings
array is that if not carefull it will be too small (smaller
than dictionary per-service in settings).