Lock connection host->mutex in connection_handler
If a timeout occurs when a client exists and is non-Idle, the timeout
is ignored. This ensures that if a timeout happens after a connection
is picked up in client_prep_connection, then it is not destroyed
(soon after).

In connection_create, move code that adds connection to host list to
the end of the creation phase.