New setting setting: pz:max_connections
Setting pz:max_connections is a limit of number of sockets to a host.
When this limit is reached, Pazpar2 will wait up to 5 seconds for a
connection to becomes available. The client will be marked Client_Error
when it can not be searched (other clients in a session may work).
If pz:max_connections is not set for a target, a value of 30 will be
used.