Indented.
Fixed: Keeping the start of the channel list as invariant through the event loop, so we dont look at newly added IOCHANnels. This could be the cause of the messed-up http show call.

Started on the refactoring of moving to yaz_poll, but holding back to see if this fix works.