New special sort parameter value 'position'.
The 'position' sorts merged records by their original position
from the remote target. This is primarily useful for debugging and
may be used for targets that already perform some kind of relevance
ranking. Note that sort by default is decreasing; so to get records
in their original order sort=position:1 must be used.