Full termlist saved (as before) and also sorted.
No max of termlist(s) any more. The memory usage is the same
(in fact less - over time), but we have a full quick sort instead
of a smaller quick sort, But NO insertion sort for each term (which
is probably similar to extra quick sort). This makes termlists
stable.