add support for a <metadata> container tag
so one can have a setup like

<metadata>
	<metadata name="language" brief="yes" merge="unique" termlist="yes"/>
	<metadata name="id" brief="yes"/>
	<metadata name="lccn" merge="unique"/>
</metadata>

The underlying desire is to be able to include the complete metadata setup with XInclude.
This requires us to have a single container with all elements.

This patch adjust the metadata element counting and evaluation to interpret both metadata tags on their own and containers with metadata tags in them.
Attributes of a metadata container tag are ignored.