Fix hang of record with esn/syntax (2nd)
Command record with esn and/or syntax makes a definite retrieval.
If that hits a cached record it would not see record events from
ZOOM and handle those. This commit ensures that connection_continue
calls event handling function non_block_events if necessary.
Since the session is already locked in some cases, we must ensure
that non_block_events do not (again) block on session.