Part of PAZ-996.
Move CC By-ND-SA to end of the list of the recognised licences, since
there is actually no such licence! (There can't be: ND prohibits the
creation of derivatives, and SA is a restriction on the licencing of
derivatives. If anything I guess this is equivclent to straight CC
By-ND, since the SA clause can never fire.)